Fixed observations_per_class datatype

+++ include/dataset.h
@@ -77,6 +77,6 @@ uint_fast32_t remove_duplicates(uint_fast64_t *dataset);
  * calculation of the disjoint matrix
  */
 void fill_class_arrays(uint_fast64_t *dataset, uint_fast32_t *n_items_per_class,
-		uint_fast32_t **observations_per_class);
+		uint_fast64_t **observations_per_class);
 
 #endif
